+++ core/src/lib.rs
@@ -33,9 +33,7 @@ pub mod util;
 pub use primary_keys::MongoPrimaryKeys;
 mod foreign_keys;
 pub use foreign_keys::MongoForeignKeys;
-pub mod load_library;
+pub mod cluster_type;
+mod mongosqltranslate;
 pub mod oidc_auth;
 pub mod test_config;
-
-pub mod cluster_type;
-pub mod mongosqltranslate_data_types;
